requirements.txt
PyQt6>=6.5.0
anthropic>=0.18.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
openai>=1.12.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
python-docx>=0.8.11
ebooklib>=0.18
google-generativeai>=0.3.0
